Add CMake and build configuration files for x86_64 architecture
- Created build_file_index.txt to track build files.
- Added cmake_install.cmake for installation configuration.
- Generated configure_fingerprint.bin for structured logging.
- Included metadata_generation_command.txt with build parameters.
- Added prefab_config.json to manage prefab settings.
- Created symbol_folder_index.txt for symbol folder tracking.
- Introduced app_keys.dart to store PayPal API credentials.

+++ lib/features/checkout/domain/entities/paypal_payment_entity/details.dart
@@ -3,10 +3,9 @@ import 'package:fruit_hub/features/checkout/domain/entities/order_entity.dart';
 class Details {
   String? subtotal;
   String? shipping;
-  int? shippingDiscount;
+  double? shippingDiscount;
 
   Details({this.subtotal, this.shipping, this.shippingDiscount});
-
 
   Map<String, dynamic> toJson() => {
         'subtotal': subtotal,
@@ -16,6 +15,6 @@ class Details {
   factory Details.fromEntity(OrderEntity orderEntity) => Details(
         subtotal: orderEntity.cartEntity.calculateTotalPrice().toString(),
         shipping: orderEntity.calculateShippingCost().toString(),
-        shippingDiscount: orderEntity.calculateShippingDiscount().toInt(),
+        shippingDiscount: orderEntity.calculateShippingDiscount(),
       );
 }
